# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.CrRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.CrResponse:
    properties:
      id:
        type: string
      username:
        type: string
    type: object
  handlers.LogResponse:
    properties:
      URL:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Provide login token to the chat service
      parameters:
      - description: Add user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CrRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.LogResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Login user
      tags:
      - users
  /users:
    get:
      description: get users
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: User List
      tags:
      - users
    post:
      consumes:
      - application/json
      description: post user
      parameters:
      - description: Login user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CrRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CrResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Creat user
      tags:
      - users
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"